Fixed bags, developed sumMuxProg

+++ Devices/alu.h
@@ -10,7 +10,7 @@ public:
     ALU();
     ~ALU();
 
-    quint32 addOp(const quint32 a, const quint16 b, bool doubleSize);
+    quint32 addOp(const quint32 a, const quint32 b, bool doubleSize);
     quint16 subOp(const quint16 a, const quint16 b);
     quint32 muxOp(const quint16 a, const quint16 b);
     quint16 andOp(const quint16 a, const quint16 b);
